requirements.txt
APScheduler==3.10.4
feedparser==6.0.11
python-dotenv==1.0.1
requests==2.32.3
deep-translator==1.11.4
beautifulsoup4==4.12.3
